fix: rename url.ts helper to platform.ts
